+++ tiled/server/simple.py
@@ -90,6 +90,7 @@ class SimpleTiledServer:
         directory: Optional[Union[str, pathlib.Path]] = None,
         api_key: Optional[str] = None,
         port: int = 0,
+        readable_storage: Optional[Union[str, pathlib.Path]] = None,
     ):
         # Delay import to avoid circular import.
         from ..catalog import from_uri as catalog_from_uri
@@ -100,9 +101,9 @@ class SimpleTiledServer:
             directory = pathlib.Path(tempfile.mkdtemp())
             self._cleanup_directory = True
         else:
-            directory = pathlib.Path(directory)
+            directory = pathlib.Path(directory).resolve()
             self._cleanup_directory = False
-        directory.mkdir(parents=True, exist_ok=True)
+        (directory / "data").mkdir(parents=True, exist_ok=True)
         # In production we use a proper 32-bit token, but for brevity we
         # use just 8 here. This server only accepts connections on localhost
         # and is not intended for production use, so we think this is an
@@ -121,8 +122,12 @@ class SimpleTiledServer:
 
         self.catalog = catalog_from_uri(
             directory / "catalog.db",
-            writable_storage=directory / "data",
+            writable_storage=[
+                directory / "data",
+                f"duckdb:///{str(directory / 'storage.duckdb')}",
+            ],
             init_if_not_exists=True,
+            readable_storage=readable_storage,
         )
         self.app = build_app(
             self.catalog, authentication={"single_user_api_key": api_key}